+++ RELEASE-NOTES.md
@@ -35,13 +35,14 @@ END TEMPLATE-->
 
 ### Breaking changes
 
-*None yet*
+* `ToolshedEnvironment` initialization  now requires that entity systems have already been initialized first.
 
 ### New features
 
 * Add a new `SerializationManager.PushComposition()` overload that takes in a single parent instead of an array of parents.
 * `BoundUserInterfaceMessageAttempt` once again gets raised as a broadcast event, in addition to being directed.
   * This effectively reverts the breaking part of the changes made in v252.0.0
+* Toolshed can now inject entity systems as dependencies
 
 ### Bugfixes